refactor: common 관련 domain 클래스 패키징
